requirement.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-jsonfield==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-jsonfield==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-jsonfield==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
